move enum to domain directory
