+++ src/command/validated_options/validated_user_inputs.rs
@@ -1,4 +1,4 @@
-use crate::validator::ambiguous_time_strategy::AmbiguousTimeStrategy;
+use crate::command::validated_options::ambiguous_time_strategy::AmbiguousTimeStrategy;
 use chrono::NaiveDateTime;
 use chrono_tz::Tz;
