+++ src/validator.rs
@@ -1,4 +1,3 @@
-pub(crate) mod ambiguous_time_strategy;
 pub(crate) mod ambiguous_time_strategy_validator;
 pub(super) mod command_options_validator;
 mod native_datetime_validator;
